bucksaltbot/requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
blinker==1.8.1
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
kombu==5.3.7
MarkupSafe==2.1.5
mwoauth==0.4.0
mwparserfromhell==0.6.6
oauthlib==3.2.2
packaging==24.0
pytest==8.3.3
prompt-toolkit==3.0.43
PyJWT==2.8.0
PyMySQL==1.1.1
pypdf==4.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywikibot==9.1.2
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
